+++ docs/releases.md
@@ -27,6 +27,7 @@ Please follow individual releases for more information.
 * Configure relationship auth rule with field instead of database key ([#2101](https://github.com/aerogear/graphback/pull/2073), fixed by [525bc9a](https://github.com/aerogear/graphback/commit/525bc9a641fa7cb1818a0727a675564e6fa12dda))
 * Fix `Could not find a declaration file for module 'bson'` error ([#2104](https://github.com/aerogear/graphback/pull/2104), fixed by [4f9ce7c](https://github.com/aerogear/graphback/commit/4f9ce7c2d6c494b33f447e1b4d6a47fbd880f353))
 * Add missing interface ([#2019](https://github.com/aerogear/graphback/pull/2109), fixed by [f46e920](https://github.com/aerogear/graphback/commit/f46e9200def565b0b0e34ccc13f7efa50f346550))
+* Generate schema subscription fields when mutations are disabled ([#2114](https://github.com/aerogear/graphback/2114), fixed by [212eb7a](https://github.com/aerogear/graphback/commit/212eb7a3e718eb102c226c237ce2448a2aa26898))
 
 ### Breaking Changes